+++ pkg/sonar/scan_test.go
@@ -1,0 +1,24 @@
+package sonar
+
+import (
+	"path/filepath"
+	"testing"
+
+	"github.com/opendevstack/pipeline/internal/projectpath"
+)
+
+func TestExtractComputeEngineTaskID(t *testing.T) {
+
+	c := testClient("")
+	want := "AVAn5RKqYwETbXvgas-I"
+	fixture := filepath.Join(projectpath.Root, "test/testdata/fixtures/sonar", ReportTaskFilename)
+	got, err := c.ExtractComputeEngineTaskID(fixture)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// check extracted status matches
+	if got != want {
+		t.Fatalf("want %s, got %s", want, got)
+	}
+}